ai-engine/requirements.txt
# Core
yfinance>=0.2.40
fredapi>=0.5.2
pandas>=2.2.0
numpy>=1.26.0

# Backtesting
bt>=1.1.0
vectorbt>=0.26.0
quantstats>=0.0.62

# Sentiment / NLP
transformers>=4.40.0
torch>=2.2.0
finbert-embedding>=0.1.0

# AI APIs
groq>=0.5.0
anthropic>=0.25.0

# Data clients
requests>=2.31.0
websocket-client>=1.7.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru>=0.7.0
